Refactoring structure of files
