+++ src/pages/Provider/components/ProviderHeader/index.js
@@ -1,0 +1,3 @@
+import ProviderHeader from './ProviderHeader';
+
+export default ProviderHeader;
