+++ src/pages/Provider/components/ProviderProduct/index.js
@@ -1,0 +1,3 @@
+import ProviderProduct from './ProviderProduct';
+
+export default ProviderProduct;
