+++ src/pages/Provider/components/ProviderProducts/index.js
@@ -1,0 +1,3 @@
+import ProviderProducts from './ProviderProducts';
+
+export default ProviderProducts;
