+++ src/pages/Provider/components/index.js
@@ -1,0 +1,5 @@
+import ProviderHeader from './ProviderHeader';
+import ProviderProduct from './ProviderProduct';
+import ProviderProducts from './ProviderProducts';
+
+export { ProviderHeader, ProviderProduct, ProviderProducts };
